Artist Delete Gig is perfect trust me

+++ gigluvver_app/views.py
@@ -274,12 +274,22 @@ def delete_gig(request):
     if request.method == 'POST':
         form = DeleteForm(request.POST)
 
-        gig_id = form['PerformerGig'].value()
+    
 
-        Gig.objects.filter(id=gig_id).delete()
+        user = get_profile(request)
+
+        gig_id = form['PerformerGig'].value()
+        
+        print("Bleh")
+        print(user.id)
+        print(form["Performers"].value()[0])
+        if int(form["Performers"].value()[0]) == user.id:
+            Gig.objects.filter(id=gig_id).delete()
+          
+        
 
         return redirect('/gigluvver_app/')
-    
+
     return render(request, 'delete_gig.html', context={'profile':get_profile(request), 'form':form})
 
 def get_profile(request):
